+++ CHANGELOG.md
@@ -11,6 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [0.2.1] - 2025-04-17
+
+### Add
+
+- Rebuild project command.
+- Deploy project command.
+
 ## [0.1.8] - 2025-04-17
 
 ### Add